feat: Add click functionality to note labels in main_window.py
This commit adds the ability to click on note labels in the main window of the application. When a note label is clicked, the `on_note_click` method is triggered. This method retrieves the color and text of the clicked note label and passes them to the `note_editor` screen. Additionally, the hexadecimal representation of the color is calculated using the `rgba_to_hex` method and displayed in the `color` field of the `note_editor` screen.